AR_django/requirements.txt
Django==4.1.7
django-cors-headers==3.14.0
django-filter==23.1
django-templated-mail==1.1.1
djangorestframework==3.14.0
djangorestframework-simplejwt==4.8.0
djoser==2.1.0
social-auth-app-django==4.0.0
social-auth-core==4.4.1
django-jazzmin==2.6.0
gunicorn==21.2.0
whitenoise==6.5.0